requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
twilio==9.3.0
google-generativeai==0.8.3
python-dotenv==1.0.1
httpx==0.27.0